Move managed nodegroup userdata creation to pkg/nodebootstrap (#3297)
